refactor: Convert session model to typescript
